issue 31 is partially addressed. the changes introduce the window length parameter into the measurement functions.
windowLength (defined in samples) determines the size of the window for computing slopes.
if the window length larger than 1 is used, a kind of smoothing and low pass filtering is applied.
If slope estimates from data with different sampling rates should be compared, the
window should be choosen in such a way that the length in milliseconds is approximately the same.
This reduces some variability, the slope estimates are more robust and comparible.

However, windowLength is currently fixed to one sample, so measurement results should be unaffected
by this patch.