Fix date and time for ABF2 files
Date and time values are correctly read from the file, but
stfio::importABF2File() does not correctly translate these to the format
expected by Recording::SetDateTime(). Time is stored in the file as
milliseconds after midnight, but the code treats it as seconds. Also, the
date is encoded with a four digit year and months from 1-12, whereas
Recording::SetDateTime() expects the number of years after 1900 and months
from 0-11.